db/requirements.txt
mysql-connector-python>=2.2.9
python-dotenv>=0.20.0 